requirements.txt
# AUTOFORGE Dependencies

# LLM Clients
google-generativeai>=0.3.0
openai>=1.0.0

# YAML/JSON processing
pyyaml>=6.0
jsonschema>=4.0.0

# Code generation & validation
jinja2>=3.0.0
black>=23.0.0
isort>=5.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Static analysis
pylint>=3.0.0

# CLI
click>=8.0.0
rich>=13.0.0

# Utilities
python-dotenv>=1.0.0
